+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Details panel responds to mouse scroll in all tabs
+- Details panel sets COLUMNS to allow jj diff tool to fit window